% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.CI.R
\name{binom.CI}
\alias{binom.CI}
\title{Confidence intervals for binomial counts or proportions}
\usage{
binom.CI(x, n, conf.level = 0.95, type = c("wilson", "exact",
  "approximate"))
}
\arguments{
\item{x}{Number of successes in n trials, can be a vector.}

\item{n}{Number of Bernoulli trials, can be a vector.}

\item{conf.level}{Confidence level (default = 0.95), can be a vector.}

\item{type}{Possible values are "wilson", "exact" and "approximate". See details.}
}
\value{
This function returns a n x 6 matrix with the following colnames:

\code{x} Number of successes in n trials

\code{n} Number of Bernoulli trials

\code{prop} Proportion = x / n

\code{lower} Lower confidence interval limit

\code{upper} Upper Confidence interval limit

\code{conf.level} Confidence level
}
\description{
Calculates confidence intervals for binomial counts or proportions
}
\details{
\code{binom.CI} with type "exact", calculates exact confidence intervals for binomial counts or proportions. This function uses R's \code{\link[stats]{binom.test}} function; however, the arguments to this function can be numeric vectors of any length.

\code{binom.CI} with type "wilson", calculates confidence intervals for binomial counts or proportions using Wilson's formula which approximate the exact method. The arguments to this function can be numeric vectors of any length.

\code{binom.CI} with type "approximate" calculates confidence intervals for binomial counts or proportions using a normal approximation to the binomial distribution. The arguments to this function can be numeric vectors of any length.

This function is a clone from epitools binomial confidence intervals functions, but with a wrapper to include the method as an option instead of separate functions.
}
\examples{
binom.CI(1:10, seq(10, 100, 10), type = "exact")
binom.CI(1:10, seq(10, 100, 10), type = "wilson")
binom.CI(1:10, seq(10, 100, 10), type = "approximate")
}
\references{
Tomas Aragon, et al. Applied Epidemiology Using R.

Kenneth Rothman (2002), Epidemiology: An Introduction, Oxford University Press, 1st Edition.
}
\seealso{
\code{\link[stats]{binom.test}}, \code{\link{SS}}, \code{\link{diagnosis}}
}

