% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.auROCc.R
\name{np.auROCc}
\alias{np.auROCc}
\title{Area under the receiver operating characteristic (ROC) curve}
\usage{
np.auROCc(ref, test, CL = 0.95, reverse = "auto")
}
\arguments{
\item{ref}{The reference standard. A column in a data frame or a vector indicating the classification by the reference test. The reference standard must be coded either as 0 (absence of the condition) or 1 (presence of the condition)}

\item{test}{The index test or test under evaluation. A column in a dataset or vector indicating the test results in a continuous scale.}

\item{CL}{Confidence limit. The limits of the confidence interval. Must be coded as a number in a range from 0 to 1. Default value is 0.95}

\item{reverse}{"auto" (default), TRUE or FALSE are the acceptable values. ROC analysis assumes that higher values of the test are from subjects with the condition, and lower values are from subjects without the condition. If it occurs the other way around, the ROC analysis and its interpretation must be reversed. If "auto", \code{np.auROCc} internally checks if the auc is below 0.5. If this is the case, it returns a warning and makes the reversion as 1 - auc.}
}
\value{
\code{AUC} The area under the ROC curve.

\code{AUC.SE} The area under the ROC curve standard error.

\code{AUC.lower.CL} The area under the ROC curve lower confidence limit.

\code{AUC.upper.CL} The area under the ROC curve lower confidence limit.
}
\description{
\code{np.auROCc} estimates the non-parametric (trapezoidal) area under the ROC curve (also known as Mann-Whitney statistic), its standard error and confidence interval (by DeLong method). Its confidence interval has the purpose to test if the AUC is different from 0.5 (the auc null hypothesis value). It is not appropriate to use this confidence interval to compare with other auc due to lack of adjustment of possible correlation.
}
\examples{

data("tutorial")

# The reference standard is not coded correctly
np.auROCc(ref = ifelse(tutorial$Gold == "pos", 1, 0), test = tutorial$Test_B)

# Artificially reversing the test to see the function reversing automatically
np.auROCc(ref = ifelse(tutorial$Gold == "pos", 0, 1), test = tutorial$Test_B)

rm(tutorial)
}
\seealso{
\code{\link{SS}}, \code{\link{TGROC}}
}
