\name{tutorial}
\docType{data}
\alias{tutorial}
\title{Example data set for continuos tests results}
\description{
This data set gives results of trhee serological tests - continuous scale - and its
correspondig reference test classfication. Also gives age as four ordinal groups.
}
\usage{tutorial}
\format{A data.frame containing 5 variables and 170 observations.}
\source{Not informed}
\references{
Not informed
}
\keyword{datasets}