% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocdata.R
\docType{data}
\name{rocdata}
\alias{rocdata}
\title{Example data set for continuos tests results}
\format{A \code{data.frame} containing 3 variables and 148 observations.

\itemize{
  \item \code{Gold} The reference standard as 0 and 1.
  \item \code{test1} Results for test 1 - numeric.
  \item \code{test2} Results for test 2 - numeric.
}}
\source{
Not informed.
}
\usage{
data(rocdata)
}
\description{
This data set gives results of two serological tests - continuous scale - and its correspondig reference test classfication - 0 for negative and 1 for positive.
}
