% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\docType{data}
\name{tutorial}
\alias{tutorial}
\title{Example data set for continuos tests results.}
\format{A \code{data.frame} containing 5 variables and 170 observations.

\itemize{
  \item \code{Gold} The reference standard as "pos" and "neg".
  \item \code{Age} Categories of age.
  \item \code{Test_A} Results for test A - numeric.
  \item \code{Test_B} Results for test B - numeric.
  \item \code{Test_C} Results for test C - numeric.
}}
\source{
This is the example dataset that came within the ancient CMDT software.
}
\usage{
data(tutorial)
}
\description{
This data set gives results of trhee serological tests - continuous scale - and its correspondig reference test classfication. Also gives age as four ordinal groups.
}
