% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\docType{data}
\name{tutorial}
\alias{tutorial}
\title{Example data set for continuos tests results.}
\format{A data.frame containing 5 variables and 170 observations.}
\source{
This is the example dataset that came within the ancient CMDT software.
}
\usage{
tutorial
}
\description{
This data set gives results of trhee serological tests - continuous scale - and its correspondig reference test classfication. Also gives age as four ordinal groups.
}
\details{
\itemize{
  \item Gold The reference standard as "pos" and "neg".
  \item Age Categories of age.
  \item Test_A Results for test A - numeric.
  \item Test_B Results for test B - numeric.
  \item Test_C Results for test C - numeric.
}
}
\keyword{datasets}

